Rewrite using RC
It seems using reference counting for the objects created by the parser
solves much of the issues wrt cloning values, copying memory
unnecessarily. Implementing this change is a lot of busy-work, but
all-in-all not very interesting.

The next step is to create an AST rewriter that rewrites all modules to
use module-prefixed names for functions and types. We can then slam them
together in a single runtime module, pruning away unused functions and
types when necessary. This also prepares for removing the interpreter
and generating (byte)code in the future.